File structure reorganized
