+++ .github/workflows/combined_tests.yml
@@ -39,7 +39,7 @@ jobs:
           go get -v -t -d ./...
           go test -v  ./...
           go build -v .
-        working-directory: 'service/'
+        working-directory: './'
 
       - name: Run Backend service
         env:
@@ -56,12 +56,4 @@ jobs:
           yarn install --frozen-lockfile
           yarn build
           yarn test:ci
-        working-directory: 'api/typescript/'
-
-      - name: Golang API step
-        run: |
-          go get -v -t -d ./...
-          go test -v  ./...
-          go build -v .
-        working-directory: 'api/go-statika/'
-
+        working-directory: 'typescript/'
