--- api/go-statika/go.mod
@@ -1,9 +1,0 @@
-module github.com/raver119/statika/api/go-statika
-
-require (
-	github.com/go-resty/resty/v2 v2.3.0
-	github.com/google/uuid v1.2.0
-	github.com/stretchr/testify v1.7.0
-)
-
-go 1.11
