+++ local_storage_test.go
