+++ typescript/.npmignore
