+++ typescript/src/Statika.ts
