+++ typescript/src/Utilities.ts
