+++ typescript/src/classes/api/AuthenticationBean.ts
