+++ typescript/src/classes/api/Communicator.ts
