+++ typescript/src/classes/entities/ApiResponse.ts
