+++ typescript/src/classes/entities/AuthenticationResponse.ts
