+++ typescript/src/classes/entities/FileEntry.ts
