+++ typescript/src/classes/entities/HttpStatusCode.ts
