+++ typescript/src/classes/entities/ListResponse.ts
