+++ typescript/src/classes/entities/UploadRequest.ts
