+++ typescript/src/classes/entities/UploadResponse.ts
