+++ typescript/src/classes/exceptions/AuthenticationException.ts
