+++ typescript/src/classes/system/EndpointsCoordinates.ts
